Rename page_sorter into page_tree
This module now handles more than sorting.
