+++ package/admin.js
@@ -1,6 +1,6 @@
 import translate from "./src/i18n"
 import NodeTree from "./src/node_tree"
-import PageSorter from "./src/page_sorter"
+import PageTree from "./src/page_tree"
 
 // Global Alchemy object
 if (typeof window.Alchemy === "undefined") {
@@ -12,5 +12,5 @@ Object.assign(Alchemy, {
   // Global utility method for translating a given string
   t: translate,
   NodeTree,
-  PageSorter
+  PageTree
 })
